Made the makeCollection and makeFullCollection methods final since set tests should use makeEmptySet and makeFullSet rather than their collection counterparts.
Added getSet() and getConfirmedSet() methods which wrap the collection and
confirmed fixtures so that tests don't need to perform the cast to a Set.

Modified tests that used makeEmptySet and makeFullSet to use the fixture
getSet() along with resetEmpty() and resetFull().  This allows for more calls
to verify() to ensure that calls to methods that should not modify a
collection don't actually modify it.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/collections/trunk@130724 13f79535-47bb-0310-9956-ffa450edef68