Enhanced TestMap to more completely test the Map contract.
Submitted by: Paul Jack ( pjack at sfaf dot org ).

Needed to change Paul's implementation slightly to get things compile properly
using JDK 1.2.  The 1.2 compiler thought that the "confirmed" field was
ambiguous in the anonymous inner class and thus needed to be qualified with a
ClassName.this.  In order to do that, the anonymous inner classes needed to be
made named inner classes.

This change exposes one test failure for MultiHashMap that appears to be caused
by its values() method not being backed by the map (bug 9573).


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/collections/trunk@130725 13f79535-47bb-0310-9956-ffa450edef68