Modified tests that used makeEmptyMap and makeFullMap to use the map field along with resetEmpty() and resetFull().  This allows for more calls to verify to ensure that calls to methods that should not modify a collection don't actually modify it.
Split the verify() method into four separate verification methods.  This way,
if a Map implementation needs to change the way verification is performed for
just one of the collection views, they only need to override one method and
won't need to reimplement all the other verifications.

Override TestMultiHashMap's verifyValues() method to retrieve a fresh copy of
the values collection before verification.  This ensures some verification
testing on the values, even though the values is not backed by the map (bug
9573)

Override a few more test methods and the verifyyMap and verifyEntrySet methods
for TestMultiHashMap since it's values are collections rather than the expected
objects.  These will need to be filled in with better implementations to
actually test the MultiMap contact.

Removed the SupportsPut marker interface.

Removed some duplicate tests (i.e. some old tests that weren't written with the
current framework in mind and just duplicate the tests that have been added).


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/collections/trunk@130726 13f79535-47bb-0310-9956-ffa450edef68