Adding in OrderedMap class. Similar functionality to a SequencedHashmap, it is decorable and able to sit on top of other Map implementations.
This is the first version to pass the unit tests, so I'm pushing it in. The
major thing I know it doesn't do is that keySet/entrySet and values iterator()
methods do not back their remove()'s onto the Map itself. The views themselves
do, just not their iterators yet.

This shows that we need unit tests for these :)


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/collections/trunk@131226 13f79535-47bb-0310-9956-ffa450edef68