Update travis to run japicmp in the main script and fix coveralls.
Anything that fails in after_success is ignored in travis reporting. The
checks must be done in the main script.

The japicmp was failing due to the lack of a jar to compare and so
coveralls was then not submitting. No coverage reports have been logged
by coveralls since June 2017.